Do not use decorators to convert volume names
These decorators brought complexity, but did not give advantages.
The only potential advantages was auto encode of guid to base 64 form.
But this does not look useful in the common case. It looks as a partial
solution of the common problem (volume name length limitation), but this
implicitly convertation brings more questions.